Import of SAMV/Dowse-BadSSH-0.05 from CPAN.
gitpan-cpan-distribution: Dowse-BadSSH
gitpan-cpan-version:      0.05
gitpan-cpan-path:         SAMV/Dowse-BadSSH-0.05.tar.gz
gitpan-cpan-author:       SAMV
gitpan-cpan-maturity:     released

+++ inc/File/Temp.pm
@@ -1,31 +1,39 @@
 #line 1
 package File::Temp;
 
-#line 122
+#line 133
 
 # 5.6.0 gives us S_IWOTH, S_IWGRP, our and auto-vivifying filehandls
-# People would like a version on 5.005 so give them what they want :-)
-use 5.005;
+# People would like a version on 5.004 so give them what they want :-)
+use 5.004;
 use strict;
 use Carp;
 use File::Spec 0.8;
 use File::Path qw/ rmtree /;
 use Fcntl 1.03;
+use IO::Seekable; # For SEEK_*
 use Errno;
 require VMS::Stdio if $^O eq 'VMS';
+
+# pre-emptively load Carp::Heavy. If we don't when we run out of file
+# handles and attempt to call croak() we get an error message telling
+# us that Carp::Heavy won't load rather than an error telling us we
+# have run out of file handles. We either preload croak() or we
+# switch the calls to croak from _gettemp() to use die.
+require Carp::Heavy;
 
 # Need the Symbol package if we are running older perl
 require Symbol if $] < 5.006;
 
 ### For the OO interface
-use base qw/ IO::Handle /;
-use overload '""' => "STRINGIFY";
-
+use base qw/ IO::Handle IO::Seekable /;
+use overload '""' => "STRINGIFY", fallback => 1;
 
 # use 'our' on v5.6.0
-use vars qw($VERSION @EXPORT_OK %EXPORT_TAGS $DEBUG);
+use vars qw($VERSION @EXPORT_OK %EXPORT_TAGS $DEBUG $KEEP_ALL);
 
 $DEBUG = 0;
+$KEEP_ALL = 0;
 
 # We are exporting functions
 
@@ -43,6 +51,10 @@ use base qw/Exporter/;
 	      mkstemps
 	      mkdtemp
 	      unlink0
+	      cleanup
+	      SEEK_SET
+              SEEK_CUR
+              SEEK_END
 		};
 
 # Groups of functions for export
@@ -50,14 +62,15 @@ use base qw/Exporter/;
 %EXPORT_TAGS = (
 		'POSIX' => [qw/ tmpnam tmpfile /],
 		'mktemp' => [qw/ mktemp mkstemp mkstemps mkdtemp/],
+		'seekable' => [qw/ SEEK_SET SEEK_CUR SEEK_END /],
 	       );
 
 # add contents of these tags to @EXPORT
-Exporter::export_tags('POSIX','mktemp');
+Exporter::export_tags('POSIX','mktemp','seekable');
 
 # Version number
 
-$VERSION = '0.14';
+$VERSION = '0.18';
 
 # This is a list of characters that can be used in random filenames
 
@@ -68,7 +81,7 @@ my @CHARS = (qw/ A B C D E F G H I J K L M N O P Q R S T U V W X Y Z
 
 # Maximum number of tries to make a temp file before failing
 
-use constant MAX_TRIES => 10;
+use constant MAX_TRIES => 1000;
 
 # Minimum number of X characters that should be in a template
 use constant MINX => 4;
@@ -89,12 +102,12 @@ use constant HIGH     => 2;
 my $OPENFLAGS = O_CREAT | O_EXCL | O_RDWR;
 
 unless ($^O eq 'MacOS') {
-  for my $oflag (qw/ FOLLOW BINARY LARGEFILE EXLOCK NOINHERIT /) {
+  for my $oflag (qw/ NOFOLLOW BINARY LARGEFILE EXLOCK NOINHERIT /) {
     my ($bit, $func) = (0, "Fcntl::O_" . $oflag);
     no strict 'refs';
     $OPENFLAGS |= $bit if eval {
       # Make sure that redefined die handlers do not cause problems
-      # eg CGI::Carp
+      # e.g. CGI::Carp
       local $SIG{__DIE__} = sub {};
       local $SIG{__WARN__} = sub {};
       $bit = &$func();
@@ -117,7 +130,7 @@ unless ($^O eq 'MacOS') {
     no strict 'refs';
     $OPENTEMPFLAGS |= $bit if eval {
       # Make sure that redefined die handlers do not cause problems
-      # eg CGI::Carp
+      # e.g. CGI::Carp
       local $SIG{__DIE__} = sub {};
       local $SIG{__WARN__} = sub {};
       $bit = &$func();
@@ -220,14 +233,14 @@ sub _gettemp {
   # Substr starts from 0
   my $start = length($template) - 1 - $options{"suffixlen"};
 
-  # Check that we have at least MINX x X (eg 'XXXX") at the end of the string
+  # Check that we have at least MINX x X (e.g. 'XXXX") at the end of the string
   # (taking suffixlen into account). Any fewer is insecure.
 
   # Do it using substr - no reason to use a pattern match since
   # we know where we are looking and what we are looking for
 
   if (substr($template, $start - MINX + 1, MINX) ne 'X' x MINX) {
-    ${$options{ErrStr}} = "The template must contain at least ".
+    ${$options{ErrStr}} = "The template must end with at least ".
       MINX . " 'X' characters\n";
     return ();
   }
@@ -299,7 +312,7 @@ sub _gettemp {
     ${$options{ErrStr}} = "Parent directory ($parent) is not a directory";
     return ();
   }
-  unless (-w _) {
+  unless (-w $parent) {
     ${$options{ErrStr}} = "Parent directory ($parent) is not writable\n";
       return ();
   }
@@ -341,35 +354,27 @@ sub _gettemp {
       #      but may have O_NOINHERIT. This may or may not be in Fcntl.
       local $^F = 2;
 
-      # Store callers umask
-      my $umask = umask();
-
-      # Set a known umask
-      umask(066);
-
       # Attempt to open the file
       my $open_success = undef;
-      if ( $^O eq 'VMS' and $options{"unlink_on_close"} ) {
+      if ( $^O eq 'VMS' and $options{"unlink_on_close"} && !$KEEP_ALL) {
         # make it auto delete on close by setting FAB$V_DLT bit
 	$fh = VMS::Stdio::vmssysopen($path, $OPENFLAGS, 0600, 'fop=dlt');
 	$open_success = $fh;
       } else {
-	my $flags = ( $options{"unlink_on_close"} ?
+	my $flags = ( ($options{"unlink_on_close"} && !$KEEP_ALL) ?
 		      $OPENTEMPFLAGS :
 		      $OPENFLAGS );
 	$open_success = sysopen($fh, $path, $flags, 0600);
       }
       if ( $open_success ) {
 
-	# Reset umask
-	umask($umask) if defined $umask;
+	# in case of odd umask force rw
+	chmod(0600, $path);
 
 	# Opened successfully - return file handle and name
 	return ($fh, $path);
 
       } else {
-	# Reset umask
-	umask($umask) if defined $umask;
 
 	# Error opening file - abort with error
 	# if the reason was anything but EEXIST
@@ -383,23 +388,13 @@ sub _gettemp {
       }
     } elsif ($options{"mkdir"}) {
 
-      # Store callers umask
-      my $umask = umask();
-
-      # Set a known umask
-      umask(066);
-
       # Open the temp directory
       if (mkdir( $path, 0700)) {
-	# created okay
-	# Reset umask
-	umask($umask) if defined $umask;
+	# in case of odd umask
+	chmod(0700, $path);
 
 	return undef, $path;
       } else {
-
-	# Reset umask
-	umask($umask) if defined $umask;
 
 	# Abort with error if the reason for failure was anything
 	# except EEXIST
@@ -505,9 +500,17 @@ sub _replace_XX {
   } else {
     $path =~ s/X(?=X*\z)/$CHARS[ int( rand( $#CHARS ) ) ]/ge;
   }
-
   return $path;
 }
+
+# Internal routine to force a temp file to be writable after
+# it is created so that we can unlink it. Windows seems to occassionally
+# force a file to be readonly when written to certain temp locations
+sub _force_writable {
+  my $file = shift;
+  chmod 0600, $file;
+}
+
 
 # internal routine to check to see if the directory is safe
 # First checks to see if the directory is not owned by the
@@ -542,11 +545,11 @@ sub _is_safe {
   return 1 if $^O eq 'VMS';  # owner delete control at file level
 
   # Check to see whether owner is neither superuser (or a system uid) nor me
-  # Use the real uid from the $< variable
+  # Use the effective uid from the $> variable
   # UID is in [4]
-  if ($info[4] > File::Temp->top_system_uid() && $info[4] != $<) {
+  if ($info[4] > File::Temp->top_system_uid() && $info[4] != $>) {
 
-    Carp::cluck(sprintf "uid=$info[4] topuid=%s \$<=$< path='$path'",
+    Carp::cluck(sprintf "uid=$info[4] topuid=%s euid=$< path='$path'",
 		File::Temp->top_system_uid());
 
     $$err_ref = "Directory owned neither by root nor the current user"
@@ -562,13 +565,13 @@ sub _is_safe {
   if (($info[2] & &Fcntl::S_IWGRP) ||   # Is group writable?
       ($info[2] & &Fcntl::S_IWOTH) ) {  # Is world writable?
     # Must be a directory
-    unless (-d _) {
+    unless (-d $path) {
       $$err_ref = "Path ($path) is not a directory"
       if ref($err_ref);
       return 0;
     }
     # Must have sticky bit set
-    unless (-k _) {
+    unless (-k $path) {
       $$err_ref = "Sticky bit not set on $path when dir is group|world writable"
 	if ref($err_ref);
       return 0;
@@ -714,34 +717,62 @@ sub _can_do_level {
 
 {
   # Will set up two lexical variables to contain all the files to be
-  # removed. One array for files, another for directories
-  # They will only exist in this block
-  # This means we only have to set up a single END block to remove all files
-  # @files_to_unlink contains an array ref with the filehandle and filename
-  my (@files_to_unlink, @dirs_to_unlink);
+  # removed. One array for files, another for directories They will
+  # only exist in this block.
+
+  #  This means we only have to set up a single END block to remove
+  #  all files. 
+
+  # in order to prevent child processes inadvertently deleting the parent
+  # temp files we use a hash to store the temp files and directories
+  # created by a particular process id.
+
+  # %files_to_unlink contains values that are references to an array of
+  # array references containing the filehandle and filename associated with
+  # the temp file.
+  my (%files_to_unlink, %dirs_to_unlink);
 
   # Set up an end block to use these arrays
   END {
-    # Files
-    foreach my $file (@files_to_unlink) {
-      # close the filehandle without checking its state
-      # in order to make real sure that this is closed
-      # if its already closed then I dont care about the answer
-      # probably a better way to do this
-      close($file->[0]);  # file handle is [0]
-
-      if (-f $file->[1]) {  # file name is [1]
-	unlink $file->[1] or warn "Error removing ".$file->[1];
-      }
-    }
-    # Dirs
-    foreach my $dir (@dirs_to_unlink) {
-      if (-d $dir) {
-	rmtree($dir, $DEBUG, 0);
-      }
-    }
-
+    cleanup();
   }
+
+  # Cleanup function. Always triggered on END but can be invoked
+  # manually.
+  sub cleanup {
+    if (!$KEEP_ALL) {
+      # Files
+      my @files = (exists $files_to_unlink{$$} ?
+		   @{ $files_to_unlink{$$} } : () );
+      foreach my $file (@files) {
+	# close the filehandle without checking its state
+	# in order to make real sure that this is closed
+	# if its already closed then I dont care about the answer
+	# probably a better way to do this
+	close($file->[0]);  # file handle is [0]
+
+	if (-f $file->[1]) {  # file name is [1]
+	  _force_writable( $file->[1] ); # for windows
+	  unlink $file->[1] or warn "Error removing ".$file->[1];
+	}
+      }
+      # Dirs
+      my @dirs = (exists $dirs_to_unlink{$$} ?
+		  @{ $dirs_to_unlink{$$} } : () );
+      foreach my $dir (@dirs) {
+	if (-d $dir) {
+	  rmtree($dir, $DEBUG, 0);
+	}
+      }
+
+      # clear the arrays
+      @{ $files_to_unlink{$$} } = ()
+	if exists $files_to_unlink{$$};
+      @{ $dirs_to_unlink{$$} } = ()
+	if exists $dirs_to_unlink{$$};
+    }
+  }
+
 
   # This is the sub called to register a file for deferred unlinking
   # This could simply store the input parameters and defer everything
@@ -766,7 +797,9 @@ sub _can_do_level {
 	# Directory exists so store it
 	# first on VMS turn []foo into [.foo] for rmtree
 	$fname = VMS::Filespec::vmspath($fname) if $^O eq 'VMS';
-	push (@dirs_to_unlink, $fname);
+	$dirs_to_unlink{$$} = [] 
+	  unless exists $dirs_to_unlink{$$};
+	push (@{ $dirs_to_unlink{$$} }, $fname);
 
       } else {
 	carp "Request to remove directory $fname could not be completed since it does not exist!\n" if $^W;
@@ -777,7 +810,9 @@ sub _can_do_level {
       if (-f $fname) {
 
 	# file exists so store handle and name for later removal
-	push(@files_to_unlink, [$fh, $fname]);
+	$files_to_unlink{$$} = []
+	  unless exists $files_to_unlink{$$};
+	push(@{ $files_to_unlink{$$} }, [$fh, $fname]);
 
       } else {
 	carp "Request to remove file $fname could not be completed since it is not there!\n" if $^W;
@@ -790,7 +825,7 @@ sub _can_do_level {
 
 }
 
-#line 947
+#line 997
 
 sub new {
   my $proto = shift;
@@ -822,14 +857,17 @@ sub new {
 
   # Store unlink information in hash slot (plus other constructor info)
   %{*$fh} = %args;
-  ${*$fh}{UNLINK} = $unlink;
 
+  # create the object
   bless $fh, $class;
+
+  # final method-based configuration
+  $fh->unlink_on_destroy( $unlink );
 
   return $fh;
 }
 
-#line 995
+#line 1048
 
 sub filename {
   my $self = shift;
@@ -841,11 +879,21 @@ sub STRINGIFY {
   return $self->filename;
 }
 
-#line 1016
+#line 1069
+
+sub unlink_on_destroy {
+  my $self = shift;
+  if (@_) {
+    ${*$self}{UNLINK} = shift;
+  }
+  return ${*$self}{UNLINK};
+}
+
+#line 1090
 
 sub DESTROY {
   my $self = shift;
-  if (${*$self}{UNLINK}) {
+  if (${*$self}{UNLINK} && !$KEEP_ALL) {
     print "# --------->   Unlinking $self\n" if $DEBUG;
 
     # The unlink1 may fail if the file has been closed
@@ -854,12 +902,13 @@ sub DESTROY {
     # do an unlink without test. Seems to be silly
     # to do this when we are trying to be careful
     # about security
+    _force_writable( $self->filename ); # for windows
     unlink1( $self, $self->filename )
       or unlink($self->filename);
   }
 }
 
-#line 1110
+#line 1193
 
 sub tempfile {
 
@@ -932,7 +981,8 @@ sub tempfile {
   # we have to indicate temporary-ness when we open the file. In general
   # we only want a true temporary file if we are returning just the
   # filehandle - if the user wants the filename they probably do not
-  # want the file to disappear as soon as they close it.
+  # want the file to disappear as soon as they close it (which may be
+  # important if they want a child process to use the file)
   # For this reason, tie unlink_on_close to the return context regardless
   # of OS.
   my $unlink_on_close = ( wantarray ? 0 : 1);
@@ -978,7 +1028,7 @@ sub tempfile {
 
 }
 
-#line 1280
+#line 1366
 
 # '
 
@@ -1076,7 +1126,7 @@ sub tempdir  {
 
 }
 
-#line 1400
+#line 1488
 
 
 
@@ -1105,7 +1155,7 @@ sub mkstemp {
 }
 
 
-#line 1441
+#line 1531
 
 sub mkstemps {
 
@@ -1135,7 +1185,7 @@ sub mkstemps {
 
 }
 
-#line 1483
+#line 1574
 
 #' # for emacs
 
@@ -1167,7 +1217,7 @@ sub mkdtemp {
 
 }
 
-#line 1524
+#line 1617
 
 sub mktemp {
 
@@ -1188,7 +1238,7 @@ sub mktemp {
   return $tmpname;
 }
 
-#line 1584
+#line 1679
 
 sub tmpnam {
 
@@ -1209,7 +1259,7 @@ sub tmpnam {
 
 }
 
-#line 1618
+#line 1715
 
 sub tmpfile {
 
@@ -1225,7 +1275,7 @@ sub tmpfile {
 
 }
 
-#line 1661
+#line 1760
 
 sub tempnam {
 
@@ -1243,7 +1293,7 @@ sub tempnam {
 
 }
 
-#line 1723
+#line 1832
 
 sub unlink0 {
 
@@ -1257,6 +1307,10 @@ sub unlink0 {
 
   # attempt remove the file (does not work on some platforms)
   if (_can_unlink_opened_file()) {
+
+    # return early (Without unlink) if we have been instructed to retain files.
+    return 1 if $KEEP_ALL;
+
     # XXX: do *not* call this on a directory; possible race
     #      resulting in recursive removal
     croak "unlink0: $path has become a directory!" if -d $path;
@@ -1281,7 +1335,7 @@ sub unlink0 {
 
 }
 
-#line 1783
+#line 1897
 
 sub cmpstat {
 
@@ -1317,7 +1371,7 @@ sub cmpstat {
   }
 
   # this is no longer a file, but may be a directory, or worse
-  unless (-f _) {
+  unless (-f $path) {
     confess "panic: $path is no longer a file: SB=@fh";
   }
 
@@ -1354,7 +1408,7 @@ sub cmpstat {
   return 1;
 }
 
-#line 1870
+#line 1990
 
 sub unlink1 {
   croak 'Usage: unlink1(filehandle, filename)'
@@ -1368,11 +1422,17 @@ sub unlink1 {
   # Close the file
   close( $fh ) or return 0;
 
+  # Make sure the file is writable (for windows)
+  _force_writable( $path );
+
+  # return early (without unlink) if we have been instructed to retain files.
+  return 1 if $KEEP_ALL;
+
   # remove the file
   return unlink($path);
 }
 
-#line 1964
+#line 2106
 
 {
   # protect from using the variable itself
@@ -1398,10 +1458,11 @@ sub unlink1 {
   }
 }
 
-#line 2011
+#line 2151
 
 {
   my $TopSystemUID = 10;
+  $TopSystemUID = 197108 if $^O eq 'interix'; # "Administrator"
   sub top_system_uid {
     my $self = shift;
     if (@_) {
@@ -1414,6 +1475,6 @@ sub unlink1 {
   }
 }
 
-#line 2084
+#line 2273
 
 1;
